Answers The answer for a given question is now added to the Chat. If the text extends 125 characters (constant value), the texts gets a "Read More?". Also added AJAX error handling upon getting HTTP500, you now get error message posted in Chat (and console.log for 404). There was also an annoying issue with the returned answer <body>.  Since it contains HTML, this lead to links being posted and overflowing to everything printed afterwards (from the chatbot). This is fixed by dual check if links should be removed and if it contains the <a> tag. If so, jQuery removes it by use of .text()
"Issues":
Answer marked as correct is only selected if its first (need to add loop to check if, and which is correct).
Need to add possibility to read more.
Need to add possibility to either see more answers, or re-phrase question (which in turn will require comparison on previous results).